+++ src/cmds/run.js
@@ -29,6 +29,16 @@ export default {
           type: 'boolean',
           describe: '',
           default: userConfig.run.bail
+        },
+        prefix: {
+          type: 'boolean',
+          describe: 'Prefix output with the package name',
+          default: userConfig.run.prefix
+        },
+        concurrency: {
+          type: 'number',
+          describe: 'How many scripts to run at the same time',
+          default: userConfig.run.concurrency
         }
       })
       .positional('script', {