+++ src/run.js
@@ -1,6 +1,6 @@
 import { execa } from 'execa'
 import kleur from 'kleur'
-import { everyMonorepoProject } from './utils.js'
+import { everyMonorepoProject, pipeOutput } from './utils.js'
 
 /**
  * @typedef {import("./types").GlobalOptions} GlobalOptions
@@ -27,14 +27,15 @@ export default {
         }
 
         console.info('') // eslint-disable-line no-console
-        console.info(kleur.grey(`${project.manifest.name} > npm run ${script} ${forwardArgs.join(' ')}`)) // eslint-disable-line no-console
+        console.info(kleur.grey(`${project.manifest.name}:`), `npm run ${script}${forwardArgs.length > 0 ? ` ${forwardArgs.join(' ')}` : ''}`) // eslint-disable-line no-console
 
         try {
-          await execa('npm', ['run', script, ...forwardArgs], {
+          const subprocess = execa('npm', ['run', script, ...forwardArgs], {
             cwd: project.dir,
-            stderr: 'inherit',
-            stdout: 'inherit'
+            stdio: ['ignore', 'pipe', 'pipe']
           })
+          pipeOutput(subprocess, project.manifest.name, ctx.prefix)
+          await subprocess
         } catch (/** @type {any} */ err) {
           if (ctx.bail !== false) {
             throw err
@@ -43,6 +44,8 @@ export default {
           console.info(kleur.red(err.stack)) // eslint-disable-line no-console
         }
       }
+    }, {
+      concurrency: ctx.concurrency
     })
   }
 }